Extract `ChatPreviewPanel` to its own file.
